Move Image to its final spot.
